comment cleanup

+++ src/interfaces/ILockManager.sol
@@ -18,7 +18,6 @@ interface ILockManager {
         uint256 creationTime;
     }
 
-
     function createLock(bytes32 fnftId, bytes calldata args) external returns (bytes32);
 
     function getLock(bytes32 lockId) external view returns (Lock memory);
@@ -27,7 +26,7 @@ interface ILockManager {
 
     function unlockFNFT(bytes32 salt, uint256 fnftId) external;
 
-    function getLockMaturity(bytes32 salt, uint fnftId) external view returns (bool);
+    function getLockMaturity(bytes32 salt, uint256 fnftId) external view returns (bool);
 
     function lockExists(bytes32 lockSalt) external view returns (bool);
 
@@ -37,5 +36,4 @@ interface ILockManager {
         uint256[] memory values,
         bytes[] memory calldatas
     ) external view returns (bool);
-
 }
